backend/game_platform/pyproject.toml
[tool.poetry]
name = "greenhouse-control"
version = "0.1.0"
description = ""
authors = ["nikosrk <nrekkas@gmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^2.0.0"
pydantic = "^2.7.0"
sqlmodel = "^0.0.18"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
